requirements.txt
# Backend API Dependencies
openai
python-dotenv
fastapi
uvicorn
python-multipart
requests
numpy

# RAG / Embeddings
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
rank-bm25==0.2.2

# Language Detection
langdetect

# TTS
# sounddevice
# soundfile
# faster-whisper
kokoro>=0.9.0
soundfile>=0.12.0
torch>=2.0.0
